Making hot end heat up before probing the bed with the tip
Refactoring - moving classes to individual files
