+++ ConfigurationPage/PrintLeveling/WizardPages/GettingThirdPointFor2PointCalibration.cs
@@ -1,0 +1,99 @@
+﻿/*
+Copyright (c) 2014, Lars Brubaker
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+The views and conclusions contained in the software and documentation are those
+of the authors and should not be interpreted as representing official policies,
+either expressed or implied, of the FreeBSD Project.
+*/
+
+using MatterHackers.Agg;
+using MatterHackers.Agg.UI;
+using MatterHackers.MatterControl.PrinterCommunication;
+using MatterHackers.VectorMath;
+using System;
+
+namespace MatterHackers.MatterControl.ConfigurationPage.PrintLeveling
+{
+	public class GettingThirdPointFor2PointCalibration : InstructionsPage
+	{
+		protected Vector3 probeStartPosition;
+		private ProbePosition probePosition;
+		protected WizardControl container;
+
+		public GettingThirdPointFor2PointCalibration(PrinterConfig printer, WizardControl container, string pageDescription, Vector3 probeStartPosition, string instructionsText, 
+			ProbePosition probePosition, ThemeConfig theme)
+			: base(printer, pageDescription, instructionsText, theme)
+		{
+			this.probeStartPosition = probeStartPosition;
+			this.probePosition = probePosition;
+			this.container = container;
+		}
+
+		private EventHandler unregisterEvents;
+
+		public override void OnClosed(ClosedEventArgs e)
+		{
+			unregisterEvents?.Invoke(this, null);
+
+			base.OnClosed(e);
+		}
+
+		public override void PageIsBecomingActive()
+		{
+			// first make sure there is no leftover FinishedProbe event
+			printer.Connection.LineReceived.UnregisterEvent(FinishedProbe, ref unregisterEvents);
+
+			var feedRates = printer.Settings.Helpers.ManualMovementSpeeds();
+
+			printer.Connection.MoveAbsolute(PrinterConnection.Axis.Z, probeStartPosition.Z, feedRates.Z);
+			printer.Connection.MoveAbsolute(probeStartPosition, feedRates.X);
+			printer.Connection.QueueLine("G30");
+			printer.Connection.LineReceived.RegisterEvent(FinishedProbe, ref unregisterEvents);
+
+			base.PageIsBecomingActive();
+
+			container.nextButton.Enabled = false;
+			container.backButton.Enabled = false;
+		}
+
+		private void FinishedProbe(object sender, EventArgs e)
+		{
+			StringEventArgs currentEvent = e as StringEventArgs;
+			if (currentEvent != null)
+			{
+				if (currentEvent.Data.Contains("endstops hit"))
+				{
+					printer.Connection.LineReceived.UnregisterEvent(FinishedProbe, ref unregisterEvents);
+					int zStringPos = currentEvent.Data.LastIndexOf("Z:");
+					string zProbeHeight = currentEvent.Data.Substring(zStringPos + 2);
+					probePosition.position = new Vector3(probeStartPosition.X, probeStartPosition.Y, double.Parse(zProbeHeight));
+					printer.Connection.MoveAbsolute(probeStartPosition, printer.Settings.Helpers.ManualMovementSpeeds().Z);
+					printer.Connection.ReadPosition();
+
+					UiThread.RunOnIdle(() => container.nextButton.OnClick(null));
+				}
+			}
+		}
+	}
+}
