+++ ConfigurationPage/PrintLeveling/WizardPages/LastPagelInstructions.cs
@@ -1,0 +1,86 @@
+﻿/*
+Copyright (c) 2014, Lars Brubaker
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+
+1. Redistributions of source code must retain the above copyright notice, this
+   list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright notice,
+   this list of conditions and the following disclaimer in the documentation
+   and/or other materials provided with the distribution.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+The views and conclusions contained in the software and documentation are those
+of the authors and should not be interpreted as representing official policies,
+either expressed or implied, of the FreeBSD Project.
+*/
+
+using MatterHackers.Agg.UI;
+using MatterHackers.MatterControl.PrinterCommunication;
+using MatterHackers.MatterControl.PrinterCommunication.Io;
+using MatterHackers.MatterControl.SlicerConfiguration;
+using System;
+using System.Collections.Generic;
+
+namespace MatterHackers.MatterControl.ConfigurationPage.PrintLeveling
+{
+	public class LastPagelInstructions : InstructionsPage
+	{
+		protected WizardControl container;
+		private List<ProbePosition> probePositions;
+
+		public LastPagelInstructions(PrinterConfig printer, WizardControl container, string pageDescription, string instructionsText, List<ProbePosition> probePositions, ThemeConfig theme)
+			: base(printer, pageDescription, instructionsText, theme)
+		{
+			this.probePositions = probePositions;
+			this.container = container;
+		}
+
+		public override void PageIsBecomingActive()
+		{
+			PrintLevelingData levelingData = printer.Settings.Helpers.GetPrintLevelingData();
+			levelingData.SampledPositions.Clear();
+
+			for (int i = 0; i < probePositions.Count; i++)
+			{
+				levelingData.SampledPositions.Add(probePositions[i].position);
+			}
+
+			levelingData.LevelingSystem = printer.Settings.GetValue<LevelingSystem>(SettingsKey.print_leveling_solution);
+			levelingData.CreationData = DateTime.Now;
+			// record the temp the bed was when we measured it (or 0 if no heated bed)
+			levelingData.BedTemperature = printer.Settings.GetValue<bool>(SettingsKey.has_heated_bed) ?
+				printer.Settings.GetValue<double>(SettingsKey.bed_temperature)
+				: 0;
+
+			// Invoke setter forcing persistence of leveling data
+			printer.Settings.Helpers.SetPrintLevelingData(levelingData, true);
+			PrintLevelingStream.AllowLeveling = true;
+			printer.Settings.Helpers.DoPrintLeveling(true);
+
+			if (printer.Settings.GetValue<bool>(SettingsKey.z_homes_to_max))
+			{
+				printer.Connection.HomeAxis(PrinterConnection.Axis.XYZ);
+			}
+
+			container.backButton.Enabled = false;
+
+			// Make sure when the wizard is done we turn off the bed heating
+			printer.Connection.TurnOffBedAndExtruders(TurnOff.AfterDelay);
+
+			base.PageIsBecomingActive();
+		}
+	}
+}
