+++ ConfigurationPage/PrintLeveling/WizardPages/SelectMaterialPage.cs
@@ -1,5 +1,5 @@
 ﻿/*
-Copyright (c) 2017, Kevin Pope, John Lewin
+Copyright (c) 2014, Lars Brubaker
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
@@ -27,27 +27,21 @@ of the authors and should not be interpreted as representing official policies,
 either expressed or implied, of the FreeBSD Project.
 */
 
-using System.Collections.Generic;
 using MatterHackers.Agg;
-using MatterHackers.Agg.UI;
 using MatterHackers.Localizations;
+using MatterHackers.MatterControl.SlicerConfiguration;
 
-namespace MatterHackers.MatterControl.SlicerConfiguration
+namespace MatterHackers.MatterControl.ConfigurationPage.PrintLeveling
 {
-	public class PresetsToolbar : FlowLayoutWidget
+	public class SelectMaterialPage : InstructionsPage
 	{
-		public PresetsToolbar(PrinterConfig printer)
+		public SelectMaterialPage(PrinterConfig printer, string pageDescription, string instructionsText, ThemeConfig theme)
+			: base(printer, pageDescription, instructionsText, theme)
 		{
-			this.HAnchor = HAnchor.Stretch;
-
-			int numberOfHeatedExtruders = printer.Settings.Helpers.NumberOfHotends();
-
-			this.AddChild(new PresetSelectorWidget(printer, "Quality".Localize(), Color.Yellow, NamedSettingsLayers.Quality));
-			this.AddChild(new GuiWidget(8, 0));
-			this.AddChild(new PresetSelectorWidget(printer, "Material".Localize(), Color.Orange, NamedSettingsLayers.Material));
-
-			this.Height = 60 * GuiWidget.DeviceScale;
+			var materialSelector = new PresetSelectorWidget(printer, "Material".Localize(), Color.Transparent, NamedSettingsLayers.Material);
+			materialSelector.BackgroundColor = Color.Transparent;
+			materialSelector.Margin = new BorderDouble(0, 0, 0, 15);
+			topToBottomControls.AddChild(materialSelector);
 		}
 	}
 }
-
