+++ Tests/MatterControl.Tests/MatterControl/SettingsParseTests.cs
@@ -160,7 +160,7 @@ namespace MatterControl.Tests.MatterControl
 				// shared temp
 				{
 					string[] settings = new string[] { SettingsKey.extruder_count, "2", SettingsKey.extruders_share_temperature, "1" };
-					Assert.AreEqual(GetProfile(settings).Helpers.NumberOfHotEnds(), 1);
+					Assert.AreEqual(GetProfile(settings).Helpers.NumberOfHotends(), 1);
 				}
 			}
 		}
